+++ discovery.go
@@ -3,8 +3,7 @@ package vshard_router //nolint:revive
 import (
 	"context"
 	"fmt"
-	"sync"
-	"sync/atomic"
+	"runtime/debug"
 	"time"
 
 	"golang.org/x/sync/errgroup"
@@ -28,68 +27,7 @@ const (
 
 // BucketDiscovery search bucket in whole cluster
 func (r *Router) BucketDiscovery(ctx context.Context, bucketID uint64) (*Replicaset, error) {
-	view := r.getConsistentView()
-
-	rs := view.routeMap[bucketID].Load()
-	if rs != nil {
-		return rs, nil
-	}
-
-	// it`s ok if in the same time we have few active searches
-	// mu per bucket is expansive
-	r.log().Infof(ctx, "Discovering bucket %d", bucketID)
-
-	idToReplicasetRef := r.getIDToReplicaset()
-
-	wg := sync.WaitGroup{}
-	wg.Add(len(idToReplicasetRef))
-
-	type result struct {
-		err error
-		rs  *Replicaset
-	}
-
-	// This works only for go 1.19 or higher. To support older versions
-	// we can use mutex + conditional compilation that checks go version.
-	// Example for conditional compilation: https://www.youtube.com/watch?v=5eQBKqVlNQg
-	var resultAtomic = atomic.Pointer[result]{}
-
-	for rsID, rs := range idToReplicasetRef {
-		go func(rs *Replicaset, rsID uuid.UUID) {
-			defer wg.Done()
-			if _, err := rs.BucketStat(ctx, bucketID); err == nil {
-				// It's ok if several replicasets return ok to bucket_stat command for the same bucketID,
-				// just pick any of them.
-				var res result
-				res.rs, res.err = r.BucketSet(bucketID, rsID)
-				resultAtomic.Store(&res)
-			}
-		}(rs, rsID)
-	}
-
-	wg.Wait()
-
-	res := resultAtomic.Load()
-	resultRs, err := res.rs, res.err
-
-	if err != nil || resultRs == nil {
-		return nil, Errors[9] // NO_ROUTE_TO_BUCKET
-	}
-	/*
-	   -- All replicasets were scanned, but a bucket was not
-	   -- found anywhere, so most likely it does not exist. It
-	   -- can be wrong, if rebalancing is in progress, and a
-	   -- bucket was found to be RECEIVING on one replicaset, and
-	   -- was not found on other replicasets (it was sent during
-	   -- discovery).
-	*/
-
-	return resultRs, nil
-}
-
-// BucketResolve resolve bucket id to replicaset
-func (r *Router) BucketResolve(ctx context.Context, bucketID uint64) (*Replicaset, error) {
-	if bucketID > r.cfg.TotalBucketCount {
+	if bucketID < 1 || r.cfg.TotalBucketCount < bucketID {
 		return nil, fmt.Errorf("bucket id is out of range: %d (total %d)", bucketID, r.cfg.TotalBucketCount)
 	}
 
@@ -100,51 +38,92 @@ func (r *Router) BucketResolve(ctx context.Context, bucketID uint64) (*Replicase
 		return rs, nil
 	}
 
-	// Replicaset removed from cluster, perform discovery
-	rs, err := r.BucketDiscovery(ctx, bucketID)
-	if err != nil {
-		return nil, err
+	// it`s ok if in the same time we have few active searches
+	r.log().Infof(ctx, "Discovering bucket %d", bucketID)
+
+	idToReplicasetRef := r.getIDToReplicaset()
+
+	type rsFuture struct {
+		rsID   uuid.UUID
+		future *tarantool.Future
 	}
 
-	return rs, nil
+	var rsFutures = make([]rsFuture, 0, len(idToReplicasetRef))
+	// Send a bunch of parallel requests
+	for rsID, rs := range idToReplicasetRef {
+		rsFutures = append(rsFutures, rsFuture{
+			rsID:   rsID,
+			future: rs.bucketStatAsync(ctx, bucketID),
+		})
+	}
+
+	for _, rsFuture := range rsFutures {
+		if _, err := bucketStatWait(rsFuture.future); err != nil {
+			// just skip, bucket seems do not belong to this replicaset
+			continue
+		}
+
+		// It's ok if several replicasets return ok to bucket_stat command for the same bucketID, just pick any of them.
+		rs, err := r.BucketSet(bucketID, rsFuture.rsID)
+		if err != nil {
+			r.log().Errorf(ctx, "BucketDiscovery: can't set rsID %v for bucketID %d: %v", rsFuture.rsID, bucketID, err)
+			return nil, Errors[9] // NO_ROUTE_TO_BUCKET
+		}
+
+		// TODO: should we release resources for unhandled futures?
+		return rs, nil
+	}
+
+	/*
+	   -- All replicasets were scanned, but a bucket was not
+	   -- found anywhere, so most likely it does not exist. It
+	   -- can be wrong, if rebalancing is in progress, and a
+	   -- bucket was found to be RECEIVING on one replicaset, and
+	   -- was not found on other replicasets (it was sent during
+	   -- discovery).
+	*/
+
+	return nil, Errors[9] // NO_ROUTE_TO_BUCKET
+}
+
+// BucketResolve resolve bucket id to replicaset
+func (r *Router) BucketResolve(ctx context.Context, bucketID uint64) (*Replicaset, error) {
+	return r.BucketDiscovery(ctx, bucketID)
 }
 
 // DiscoveryHandleBuckets arrange downloaded buckets to the route map so as they reference a given replicaset.
 func (r *Router) DiscoveryHandleBuckets(ctx context.Context, rs *Replicaset, buckets []uint64) {
 	view := r.getConsistentView()
-
-	count := rs.bucketCount.Load()
-
-	affected := make(map[*Replicaset]int)
+	removedFrom := make(map[*Replicaset]int)
 
 	for _, bucketID := range buckets {
 		oldRs := view.routeMap[bucketID].Swap(rs)
 
-		if oldRs != rs {
-			count++
+		if oldRs == rs {
+			continue
+		}
 
-			if oldRs != nil {
-				if _, exists := affected[oldRs]; !exists {
-					affected[oldRs] = int(oldRs.bucketCount.Load())
-				}
+		if oldRs == nil {
+			view.knownBucketCount.Add(1)
+		}
 
-				oldRs.bucketCount.Add(-1)
-			} else {
-				//                 router.known_bucket_count = router.known_bucket_count + 1
-				view.knownBucketCount.Add(1)
-			}
+		// We don't check oldRs for nil here, because it's a valid key too (if rs == nil, it means removed from unknown buckets set)
+		removedFrom[oldRs]++
+	}
+
+	var addedToRs int
+	for rs, removedFromRs := range removedFrom {
+		addedToRs += removedFromRs
+
+		switch rs {
+		case nil:
+			r.log().Debugf(ctx, "Added new %d buckets to the cluster map", removedFromRs)
+		default:
+			r.log().Debugf(ctx, "Removed %d buckets from replicaset %s", removedFromRs, rs.info.Name)
 		}
 	}
 
-	if count != rs.bucketCount.Load() {
-		r.log().Infof(ctx, "Updated %s buckets: was %d, became %d", rs.info.Name, rs.bucketCount.Load(), count)
-	}
-
-	rs.bucketCount.Store(count)
-
-	for rs, oldBucketCount := range affected {
-		r.log().Infof(ctx, "Affected buckets of %s: was %d, became %d", rs.info.Name, oldBucketCount, rs.bucketCount.Load())
-	}
+	r.log().Infof(ctx, "Added %d buckets to replicaset %s", addedToRs, rs.info.Name)
 }
 
 func (r *Router) DiscoveryAllBuckets(ctx context.Context) error {
@@ -213,31 +192,49 @@ func (r *Router) DiscoveryAllBuckets(ctx context.Context) error {
 	return nil
 }
 
-// startCronDiscovery is discovery_service_f analog with goroutines instead fibers
-func (r *Router) startCronDiscovery(ctx context.Context) error {
+// cronDiscovery is discovery_service_f analog with goroutines instead fibers
+func (r *Router) cronDiscovery(ctx context.Context) {
+	var iterationCount uint64
+
 	for {
 		select {
 		case <-ctx.Done():
 			r.metrics().CronDiscoveryEvent(false, 0, "ctx-cancel")
-
-			return ctx.Err()
+			r.log().Infof(ctx, "[DISCOVERY] cron discovery has been stopped after %d iterations", iterationCount)
+			return
 		case <-time.After(r.cfg.DiscoveryTimeout):
-			r.log().Debugf(ctx, "started new cron discovery")
+			iterationCount++
+		}
+
+		// Since the current for loop should not stop until ctx->Done() event fires,
+		// we should be able to continue execution even a panic occures.
+		// Therefore, we should wrap everyting into anonymous function that recovers after panic.
+		// (Similar to pcall in lua/tarantool)
+		func() {
+			defer func() {
+				if recovered := recover(); recovered != nil {
+					// Another one panic may happen due to log function below (e.g. bug in log().Errorf), in this case we have two options:
+					// 1. recover again and log nothing: panic will be muted and lost
+					// 2. don't try to recover, we hope that the second panic will be logged somehow by go runtime
+					// So, we choose the second behavior
+					r.log().Errorf(ctx, "[DISCOVERY] something unexpected has happened in cronDiscovery(%d): panic %v, stackstrace: %s",
+						iterationCount, recovered, string(debug.Stack()))
+				}
+			}()
+
+			r.log().Infof(ctx, "[DISCOVERY] started cron discovery iteration %d", iterationCount)
 
 			tStartDiscovery := time.Now()
 
-			defer func() {
-				r.log().Infof(ctx, "discovery done since %s", time.Since(tStartDiscovery))
-			}()
-
-			err := r.DiscoveryAllBuckets(ctx)
-			if err != nil {
+			if err := r.DiscoveryAllBuckets(ctx); err != nil {
 				r.metrics().CronDiscoveryEvent(false, time.Since(tStartDiscovery), "discovery-error")
-
-				r.log().Errorf(ctx, "cant do cron discovery with error: %s", err)
+				r.log().Errorf(ctx, "[DISCOVERY] cant do cron discovery iteration %d with error: %s", iterationCount, err)
+				return
 			}
 
+			r.log().Infof(ctx, "[DISCOVERY] finished cron discovery iteration %d", iterationCount)
+
 			r.metrics().CronDiscoveryEvent(true, time.Since(tStartDiscovery), "ok")
-		}
+		}()
 	}
 }